refactor(ProductRequestDTO, ProductResponseDTO, ProductUpdateDTO): altera tipo de 'unitValue' de Double para BigDecimal

+++ src/main/java/com/jhops10/canteen_manager/dto/product/ProductResponseDTO.java
@@ -2,10 +2,12 @@ package com.jhops10.canteen_manager.dto.product;
 
 import com.jhops10.canteen_manager.model.Product;
 
+import java.math.BigDecimal;
+
 public record ProductResponseDTO(
         Long id,
         String productName,
-        Double unitValue
+        BigDecimal unitValue
 ) {
     public static ProductResponseDTO fromEntity(Product product) {
         return new ProductResponseDTO(
